To avoid overloading of terms between "payment transfers" (transfers done upon request) and transfers - done from users who simply wants to send funds to one-another - Let's rename once again the models.
 - PaymentOrder: the entity that controls payment requests
 - PaymentEvent: A payment event is a record of the state change of a payment order
 - Payment: transfers done via blockchain/raiden/etc that related to a payment order

With these nomenclature, we can have the concept of Transfer exclusively
referring to an user that wants to initiate a transfer of tokens to
another user of the hub or EOA or raiden address.

Unify BlockchainTransfer and RaidenTransfer into ExternalTransfer, as they both should follow the same mechanism

Add model `Reserve` for control of funds that are destined for transfers.
Add signal-based flow for execution of transfers

Whenever a `Transfer` is created, a `Reserve` object with matching amount
is also created and the transfer is marked as `scheduled`.

When a transfer is executed, the total reserved can be checked and if all
conditions are satisfied, it can go through. After execution, the reserve
entry is destroyed to indicate that the balance is consumed in case of successful
 or returned to the owner in case of failed transfer.

Transfers are scheduled via signals and then executed via celery

Refactored raiden and ethereum modules into separate django apps. Sketched out functionality for external transfers
Add the concept of UserReserve (and TransferUserReserve) to control for funds that are in-flight and proper signal handling for executed/confirmed transfers
Add models to keep track of unconfirmed blockchain transfers / raiden payments
Refactor accounting models to not be dependent on web3.

Decouple Wallet model from ethereum account to allow pluggable account backends and make easier testing