Add setup of Store & Checkout model
With the store model, users may setup specific payment flows for websites.

Upon creating a store, it will have an API key and a RSA pair. The API key
might be used for authenticating requests done from the websites, while the
RSA pair is used so that our payment gateway instance can issue secure
JSON Web Tokens, which can be given to the client application. The client
applcation can then send this JWT to the store website and the store site
can securely verify the status of the payment (received, confirmed, etc)

Cleaned up command that syncs blockchain, so that it respects STARTING_BLOCK_NUMBER
and avoid loading too much of old chain history

Add views/serializers for Checkout models and added a permission class that
allows clients to access a checkout endpoint if they are using the api key
and/or authenticated as the user who owns the store

Introduce blockchain_payment_sent signal and command that track blockchain transfers

During the payment flow, we have signals to indicate when a payment has been received
and when it has been confirmed. However, this is not enough for the use case of a payment
gateway when the user initiates a blockchain transfer.

For this reason, the blockchain_payment_sent signal is introduced. with it, parts of the
application that might be interested not in the transaction itself but in the detection of
the payment (i.e, part of the checkout payment flow) can receive this signal and do whatever
it wants with the information.

A `track_pending_payments` django command is also added that listens to every pending
transaction where the recipient is any of the existing account addresses that are reserved
for payments.

Started using django channels by introduction of CheckoutConsumer, for users that want to
be updated in real-time about the status of a checkout. On any payment-relevant (such as
payment_sent, payment_received, payment_confirmed), the endpoint sends a message containing
the name of the event and a signed jwt representing the status of the checkout

Add pytest-based tests for channels