Exchange Rates via Coingecko + Refactor of docker environment files
The order of environment variable precedence is breaking the idea of
having a overrideable docker-compose and envfiles. The current solution is to
have the default ".env" file to contain all of the variables. This way, one should
define all of the desired environment variables in one file and symlink it to .env,
which can then be picked up by docker-compose and start time.

Add celery task to fetch exchange rates from coingecko and REST API endpoint to present the rate of a token pair