Handling of event data to ensure that transactions related to ERC20 tokens are indeed successful transfers.
This commit introduces the model for TransactionLog and fixes the issue that every
transaction related to an ERC20 token address was assumed to be a transfer function call.

It is still not perfect, as still need to decode the transfer event properly and none of
the web3 libraries seem to be do that in a straightforward method. The only check it makes
so far:

 - Every transfer transaction needs to have one event
 - The event must have 3 topics (transfer, from, to)
 - The event data contains the amount transferred, and it must match the parsed transaction.data