Application clean, django-update, app settings refactor
AttriuteDict (by way of the rootpath module) has a serious bug in that it sets
the value of `None` when the code is installed in a virtualenv but the working
directory is not on the sys.path. This was causing the celery job to fail to start
when django.setup was called.

The cleaner solution is to simply remove the usage of AttributeDict, so this was
an opportunity to refactor hub20.apps.core.app_settings module.

This commit also removes the apps that were not used (django-invitations, qr code)
which allows up to upgrade to django3.