Refactor of PaymentOrder/Checkout models.
Orders/checkout can only be deleted by its creator, and only when no
payments have been done.

We can use the state of the payment routes to determine if an order is
expired or not. This is done by ensuring that any route has some
method to indicate its expiration (block height for blockchain
routes, time for raiden ones) and that by establishing that an order
is expired if all of its routes are expired as well

This also adds methods to select orders that are paid or
unpaid (if the sum of payments done through the routes is bigger than
the order amount, then it is considered paid)

Add status field to PaymentOrderSerializer
Add expiration_time to RaidenPaymentRouteSerializer
Add messages for checkout events (new block added, route expired)