Implementation of dynamically loaded "Web3Indexer" classes for web3 sync
To avoid depending on syncing all transaction data obtained on the eth
node, we introduce the concept of a "Web3Indexer" classes which can be
called and run any kind of queries to the blockchain (via web3 client)
and record whatever Transaction and Block data it may be interested in.

In our case, we are mostly interested in keeping track of transactions
that involved any of the addresses of our custodial accounts. These
policies are implemented in the ethereum_money/policies/indexing
module.

For customization, a setting HUB20["WEB3_INDEXERS"] on
settings.py can be defined and it should take a list of dotted paths
to the class that are to be loaded. The command
run_blockchain_indexers then creates one process for each indexer