Change Transaction model. Add web3 mock responses
To make a transaction, some of its relevant data comes from web3
transaction receipt (logs, gas used) and some from the transaction
data (input, value). To make it simpler, the transaction model needs
to have both these parameters provided to Transaction.make

Based on factory for models, this commit introduces factories that
generate web3 data. With those we can then write unit tests for the
client functions.

Ideally, these would be the basis for a "fake" web3 provider that
could then be used to replace the HTTPProvider used in production.