Add command to record a transaction
For the case where we want to manually record one or more transactions in the
database, the command record_transactions can help with that. Simply
pass the hex strings representing the transactions you want to save
and it will be recorded (provided that it can get the data and receipt
from the blockchain)

This is useful if you are using a wallet that you have already
generated and would like to add the records from it. It can also help
the commands that track token transfers - just pass the first token
transfer that you are interested in, and the event listener will be
able to start the scan from that transaction instead of the genesis block.