Add event listener to get all txs related to current accounts
Requests all block data from the blockchain and saves transactions
where the `to` or `from` addresses are the same as any of our
accounts.

Event listener has not been added by default because it does take
quite a lot of time to run and is only effective for recording ETH
transfers.