Use pip-compile to manage dependencies and clean up some modules
Introducing raiden as a dependency revealed an extraordinary number of
dependencies that were conflicting. This commit brings a
requirements.in file that defines all dependencies and we start using
pip-compile to solve all constraints.

Some packages (ethtoken and attributedict) were causing too much
trouble and it was easier to remove them and refactor the provided
functionality instead of relying on them. ethtoken was only being used
for the EIP20_ABI definition and attributedict could be substituted by
a similar data structure provided by web3