Start raiden process based on data stored on hub20
First step of integration is to get the a Raiden client started with
the configuration parameters stored on hub20. More specifically, we
have:

 - Account info: the private key is stored on the database - part of
   the KeystoreAccount model.
 - ethereum node: the same one that is used by hub20

Raiden makes a lot of assumptions about how it is run and operated. It
expects to be run called via main and never embedded from a different
process (due to the required gevent monkey patching?). To deal with
this, we call raiden via subprocess.call and pass all the
configuration as part of the environment.

There is also the issue that Raiden currently assumes that the private
key from the account used comes via a password-protected keystore
file. To overcome this, the process that starts raiden writes two
temporary files with the required data.