Add service deposit
For Raiden to run, it requires a minimum amount of service
tokens (RDN) to be deposited in a contract that can be accessed by
the Raiden Service Bundle.

This commit brings the whole functionality that allows this deposit
to be executed. It assumes that the account used by Raiden already has
the service tokens. An API endpoint is added that lets admins make a
request for a deposit of an specified amount. This saves a
`ServiceDeposit` model and a signal handler takes any `ServiceDeposit`
and initiates the `send_service_deposit` celery task.