API for raiden operations
Adds endpoints for listing token networks and channels, as well as
endpoints for depositing and withdrawing for channels.

Each operation on the API simply creates a "RaidenManagementOrder"
entry, which should then on creation be scheduled as a background task
that executes the actual operation.

TODO: If any blockchain transaction comes out of this operation, we
should detect them and save in the RaidenManagementOrderResult entry