Change method of generated passphrases for temporary keystore files
To run raiden, we need to generate a keystore file with the private
key from the account that we want to use.

The first method used a randomly generated passphrase every time. This
works but we might end up with multiple keystore files if the process
fails and restarts.

To remediate that and still keep some things secure, the passphrase is
now generated as a SHA256 hash of django's SECRET KEY.