Refactor Transfer method, implementation of Raiden transfer
The main goal of the implementation is to place as much of the
functionality related to interacting with external systems into the
corresponding `client` modules.

In that spirit, this commit adds the implementation transfers via
RaidenClient and EthereumClient classes. The ExternalTransfer model
then defines what is the best method to complete the
transfer (currently, it assumes Raiden will always be used if the
destination address has an open channel on the TokenNetwork)

The idea of StatusModels to control the state of the transfer was
dropped in favor of Execution/Confirmation/Cancelation/Failure which
are more directly related to the flow and that are more easily queried
for the handler methods.

Testing:

Still a lot of work to do to ensure that all unit tests do not rely on
web3 proper, but this commit already introduces a proper
`Web3Mock` object that can be used to test functionality that relies
on client modules.

REST API: Changes in tokens

Token API Endpoints need to use their address for representation

The code is not unique, so to avoid any possibility of confusion we
must use the address to represent a token resourse