(fix) remove vuex getters from websocket-based event handlers
Apparently websocket objects need to be bound to the window object to
be manipulated and the vuex getters work on a different
context. Considering the getters are not needed, we can safely remove
them for now.

Later on we can establish the proper method to bind them to the window
and avoid the 'Illegal invocation' error