Simplified logic for deposit modal and transfer modal
The logic of the what current modal is being presented should be part
of the component, not the store. This also lets us keep related
functionality in the same component, instead of having to communicate
the event of a select deposit to the main app via event handling.

Transfer modal follows similar logic