Querysets and API endpoint for summarized accounting booking reports
Every model that derives from DoubleEntryAccounting model can keep
records of its own debits and classes. For the purposes of auditing
funds, it will be important to provide via the API a quick summary of
what is the current standing of each of these accounts.

The current endpoint provides a summary information of the total
credit / debit / balance of every token accounted for users,
treasury (site funds), ethereum wallets, raiden accounts (on chain and
off-chain), as well as external addresses.