(feat) Raiden setup based on its private key account
Having the system able to generate its Raiden instance can be a big
source of issues, given that the hub may not make any distinction
between accounts generated from seed phrases and private keys that are
provided by the user.

So, to avoid this issue and to have some symmetry between the
different use-cases, the settings will require that the private key to
be used for Raiden to be provided, much like the seed phrase are
required for ethereum accounts.